Fix child box handling
Set the consumed_bytes value in the Box constructor. This value is
used to parse child boxes of a container.

Derived boxes can now set the has_children flag from their parse
function to force parse_children to be called from the Box constructor
even if the constructor was originally invoked with container=False